Load the external scripts
Now that we have all the HTML & CSS in place, we need to tell the
browswer where to look to load the JavaScript files that we'll be using.
This is done with the `<script>` tag.

For this project, we'll be using two different JavaScript sources.
First, we'll load the jQuery, a JavaScript library that helps out
tremendously with adding event listeneres, manipulating the DOM, and
doing other routine tasks. jQuery is a library written and hosted by
someone else: this means we need to go grap an external link to it. To
find it, search the Web for "Google jQuery CDN". (A CDN, or content
delivery network, is basically a super-fast network designed for hosting
commonly-requested files.) On the first result page, you should see a
large list of Google hosted libraries. Find the one titled JavaScript,
copy the `<script>` tag under it, and paste it into the file `index.html`.

Now that we've loaded our helper JavaScript library, we need to load the
file which will contain our core application code. We'll be calling this
file `main.js`, so add a `<script>` tag that looks similar to the one
you just copied, but with a source of `src="main.js"` instead:

```html
<script src="main.js"></script>
```

With these lines in place in `index.html`, we should be good to go! Now
we can start writing the core JavaScript code for our game.