Add starter code in main.js
We're finally at the point where we can add some code to our `main.js`
file!

With the first few lines, we get the opportunity to see one of the many
quirks about JavaScript: something that's not really necessary but that
makes life a little bit easier.

We're going to add the following lines to our JavaScript file:

```javascript
;(function() {

})();
```

What these lines do is _terminate any previous, dangling JavaScript
statement_ (with the first semi-colon), _create an anonymous function_
(with the `function` keyword), and then _immediately call it_ (with the
`()` towards the end). This ensures that we have our own execution
context for all the code that we write. You can read more about it
[here](http://sarfraznawaz.wordpress.com/2012/01/26/javascript-self-invoking-functions/),
but you're almost always going to want to follow this convention when
writing JavaScript files.