Listen for ready event
If you're familiar with a language like C/C++ or Java, you'll know that
every application must have an entry point that's usually called `main`.
In Python there's a similar concept using globally defined variables
like `__name__`.

JavaScript has no such entry point. When JavaScript is run in a
browswer, each line is run as it is encountered. While this is powerful,
there are many times when being able to run a particular command when
the page is ready for it to be run is useful. For this purpose, we use
_event listeners_ to run a particular function once a corresponding
event occurs. Events are generated for many different actions and
conditions within the browser; one of these is the `ready` event, which
is fired on the `document` object whenever the page is, well, ready for
us.

We can then attach a _listener_, or a specialized function, to this
event, effectively allowing us to run a particular piece of code
whenever the event fires. This function can either have a name or be
anonymous. We'll chose the latter:

```javascript
$(document).ready(function() {

});
```

What this does is attach an empty function to the `ready` event of the
`document` element. This syntax is actually a call to a jQuery function:
the `$` is the variable name for all the things that jQuery provides,
and the `.ready()` method is a utility method for attaching listeners to
`ready` events.

This concept that we've used here of attaching a function to an event
and having it be run at a later point is _incredibly common_ in
JavaScript. In fact, it might be JavaScript's most powerful feature.
When used in this manner, the function that we attached to this event is
called a _callback_.