Added `play` function
Now that the "ready" event listener is in place, we can add some
boilerplate functions for where we'll write the brunt of our game code.

First, we'll need a function called `play`, which for right now will
remain empty. This function is also going to be an event listener, so
it'll take a single argument representing the event that triggered the
function to be called. You may notice that the anonymous function we
registered on the `ready` event also is an event listener, and so it
receives an event object as well. However, since we don't use it (and
almost always won't on the `ready` event), we omit it for simplicity.
This is possible because in JavaScript, _the number of arguments passed
to a function does not have to equal the number declared_.  For more
information, see [here][arguments].

Now that we've got our empty `play` function in place, we're going to
add it as an event listener on the `click` event of the submit button.
This button has an `id` of `submit`, so we can use jQuery to select that
element and bind the click event on that element to our `play` function.

[arguments]: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Functions_and_function_scope/arguments