Get the user's input for rows and columns
So now we have a function that will be executed every time the user
clicks on our submit button. Let's start filling it in with something
useful.

We said one of the things we'd have to do in order to set up the game
state was to let the user choose how many rows and columns there should
be. Using the two text that are present in the HTML of the page, we can
do just this.

By the time the user clicks on our submit button and fires off the
`play` function, the text inputs with id's of `rows` and `cols` will
be holding values corresponding to the number of rows and columns that
our board should have. We can use jQuery to extract these values, using
the `.val()` function. We'll go ahead and store these values in global
variables.

Now is also a good time to deal with the circumstance when the user
accidentally hits submit twice. What should happen? Should we start a
fresh game? Add or remove rows from the game in progress? There are
many ways we could handle this situation. For sake of simplicity, we're
just going to remove the submit button and both text inputs entirely,
thereby ensuring that we don't have to deal with this weird
cirtumstance. After selecting the form with id `args` that wraps the
text inputs and submit button, we can use the `.detach()` method to
remove them from the DOM.