Prevent form submission event from bubbling up
Once make the changes introduced in the previous step, go ahead and see
if clicking the button removes the `#args` form. (The best way to
preview your work is to use a simple Python webserver running in your
project's directory, then opening http://localhost:8000 in your browser.
If you need help with this, flag down a mentor! Otherwise, you should
also be able to simply open the index.html file in your browser.)

What you'll probably notice is that the form didn't disappear, and it
looked like the website reloaded the page. If you look carefully, you'll
probably also see that the URL changed to something with variable names,
question marks, and ampersands. This happened because of something
called `bubbling` which is another JavaScript feature which can seem
like a quirk at times.

Whenever there is a button inside a form and the button is clicked, the
browser thinks that the values of the inputs within that form need to be
submitted. If you don't specify where to submit this information on the
`<form>` tag, then it assumes that you want it to be submitted to the
current URL, effectively reloading the page.

This is pretty annoying, because we'd actually like to run our own code
instead of having the browser do something for us. To turn this feature
off, we have to understand what's actually going on. Nearly every action
a user has with a page generates some sort of event. Even though an
event might have a particular element from which it originated, the
event it self propagates up through the DOM tree from that element to
it's parent element all the way until it reaches the `<html>` tag. This
is what's known as `bubbling`, because the event bubbles up the DOM.

Our particular event handler is listening on the button itself, whereas
the form submission event is attached to the form element; we want to
shut off the bubbling before it can get there. To do this, we can use
the event argument `e` that will be passed into our `play` function. By
calling `e.preventDefault()`, we can stop the "default" action (the
event bubbling upwards) from occuring.

Try adding this line and see what happens. You should see that the
`#args` form detaches and stays detached. If not, flag down a mentor!

By now, you're probably starting to understand that there are a lot of
idiosyncrasies in JavaScript who's solutions are not obvious. In cases
like these, your best debugging tool is the ability to carefully
describe what problem you're having and searching for it on Google. At
least when you're starting out, it's likely the problems you will
experience will have relatively simple, though non-intuitive, solutions.
Being able to Google effectively is invaluable in times like these.