Initialize the game's starting state
Now that we can accept the user's input with our form, we can actually
begin setting up the initial board configuration. Specifically, the user
has told us how many rows and columns to use, so we can use a nested for
loop to set up the grid with the right dimensions.

HTML elements are display and drawn across and then down, we're going to
draw one row at a time (as opposed to one column at a time). To make
things easier, we're going to wrap each row in a div with the class
`tile-row` so that we can access arbitrary rows.

To create a new DOM element (basically, an element in the page), we use
the jQuery function (`$`) with a string argument representing the HTML
that we should use to construct that object. We just need to create a
`<div>` with a class of `tile-row`, so the string we want is `'<div
class="tile"></div>'`. We can then store this in a variable named
`$curTileRow` (note that `$` is a perfectly valid character to use in a
JavaScript variable name). Finally, we use the jQuery append method to
append this tile row to the tile wrapper.

We'll be using a multidimensional array to maintain what color each
board piece is. JavaScript arrays aren't fixed in length (because
they're actually just special objects), so we'll be dynamically adding a
row every time we need a new one. What that means for us is that within
the outer loop, where code will be executed once per row, we should add
a row to our multidimentional array. This is done with the code
`boardColors[curRow] = [];`, which sets the contents of the curRow'th
row to an empty array.

Now that we've done the required setup for each row, we can work on
adding the individual tiles that will ultimately constitute the columns.
We'll add a for loop inside our outer for loop that ranges over the
entire number of columns.

The first thing this inner loop should do is create the DOM element for
the tile residing in that column. This is done with the jQuery append
method that we used before to add a row to the outer, wrapper div. Next,
since all but the top left tiles should start out blue, we'll set the
color of this tile to blue (and worry about the one green one later).
I'll use a helper function called `setColor` to set the color of a tile.
It'll take a row, column, and a string representing a color. We'll write
this function after the next step. For now, just know that it takes care of
setting the color for us.

Finally, we'll have to set the click event on the div to handle what
should happen when the user "makes a move," or attempts to change the
color of a tile. This part is actually very sophisticated! We're going
to be using a technique called closures, which is a very powerful
concept that is used quite frequently in JavaScript.

The basic paradigm stems from the fact that JavaScript maintains a
context of all the variables and their values for every function that is
created. That means that we can create special contexts that bind
additional variables that a function would find useful to use.

First you'll note the helper function `getTile`. We'll write this helper
function in after next step as well, for but for now just now that it's
basically a wrapper around the `$` function that we've been using this
whole time.  What that means is that it returns to us the same jQuery
DOM object that we would normally get back, so we can set the `.click`
attribute of that result to bind a function to the click event of that
particular DOM element.

Next, we have an anonymous function wrapped in parentheses,
(function(r, c) {...}), followed by `(curRow, curCol)`. When you see
this, a function definition wrapped in parentheses immediately followed
by parentheses, this will create an anonymous function and _immediately
call it_ with the arguments specified in the second set of parentheses.
This is called a self executing function, and you can read more about it
[here](http://markdalgleish.com/2011/03/self-executing-anonymous-functions/).

Remember that the argument to the `.click` function needs to be a
function which takes one variable: the event object `e` that triggered
the click event. We're actually going to use this self-executing
function to _return a different function_. What's special about this
returned function is that it will have access to the parameters of the
outer function, namely the current row and current column. We can then
reference those inside our returned function; in our case, we're going
to pass them on to `move`, a third helper function that will take care
of the brunt of the game logic for us. We'll start writing that function
in three steps.

Since the function we pass to click must take an event object as an
argument, we'll make the function we return take just that argument, and
we'll pass it on to `move` just in case we need it.

That's it for the processing we need to do within the loop! If you
remember from before, I said that we'd take care of changing the upper
left tile to green later, so let's go ahead and do that now. Remember
that we use a helper function for this before, so we'll use the same
function again now.