Set the initial values on the "scoreboard"
One last thing to take care of before we start going nuts with our game
is to set up number of greens, blues, and total tiles, and then display
these values.

We'll initialize a few global variables called `reds`, `greens`, and
`blues` to 0, set `greens` to 1 (for the upper left tile), and set
`blues` to the remaining number of tiles. Obviously, the total number of
tiles is `rows * columns`, and this will be important because we will
know the user has won if the number of reds is equal to this value.

Finally, we'll use a couple more jQuery calls to display these values.
The `.html` function on a jQuery object will set the inner HTML of that
attribute to the text specified (if you pass something that's not a
string, it will try and cast it to a string). That means we can use this
method, passing it `blues` and `greens` respectively, to set these
colors to the right values. Remember that the HTML elements with id's
`blues` and `greens` house the data we want to display. Note that we
don't have to worry about setting the number of reds because it should
already be set to 0 in the HTML (you can verify this).