Define a couple useful helper functions
As I mentioned two steps ago, we're going to now define the helper
function we used to both set a tile's color given its indices, and to
get the DOM node of an element given its indices.

The first of these, `getTile`, is a one-liner, though it is a bit
complicated. We'll be using the jQuery/CSS `:nth()` pseudo-selector,
which basically allows us to access the nth element of a particular
class. We'll also be using the `>` selector, which selects the children
of a particular node. Remember that we wrapped every row in a div with
the class `tile-row`, so to get the nth row we can use `.tile-row:nth('
+ row + ')`. Then we want to access an individual tile (an element with
the class `tile`), which is a child of a `tile-row`. That means the next
selector needs to be a `>`. Finally, we want the nth tile, so in a
similar fashion to before, we want to use `.tile:nth(' + col + ')`. If
you hadn't already figured it out, the `:nth()` pseudo-selector takes as
an argument an integer which represents the index of which element it
should select.

Next up is `setColor`. We have two things to do: update our internal
representation of the colors of the tiles, and then actually change the
color of the element. Remember that we've been storing our internal
representation of the board's colors in a multidimensional array called
`boardColors`. We want to access the element at `[row][col]` and set
it's contents to `color`. We'll use this stored value later in
processing what to do when the user clicks on a tile.

The second thing to do is change the actual color of the tile. In the
CSS stylesheet (styles.css), there are a bunch of handy classes that
will turn our element a particular color depending on which class is
applied. First, we want to get the DOM element represented by the row
and column we are dealing with, which is simple enough because we have a
helper function to deal with that! We'll next use the `.removeClass`
jQuery method to remove any previous color class that had previously
been attached to this element, and then the `.addClass` method to add
the class corresponding to the color we want to make this tile.

There are a couple of things to note here. The first is that if
`removeClass` doesn't find the class you specified to remove, it doesn't
do anything.  The second is that you can _chain jQuery calls_. Note that
there are no semicolons after each line, only the last line. Every time
we call a jQuery method that's operating on a DOM element like this, we
can chain as many calls as we want on top of each other to keep
modifying the current jQuery DOM element.