Start writing move function: get adjacent tiles
We're approaching the home stretch! `move` is the last function we'll
need to implement.

Recall that `move` will be called every time someone clicks a `tile`
element, and it will have access to the current row and column because
of the closure we used when initializing the click event. We'll use this
method to handle processing all the rules of the game that we outlined
in the Overview.

In the rules, we specified that the only tiles we can actually click on
are the green ones, so let's check if the current tile is green.

Remember that we have access to the row and column of the clicked tile
from the parameters to the `move` function, and we have a
multidimensional array which will give us the color of a tile given a
row and a column. Once we have this stored in a variable, we can check
to see whether that color is `'green'`.

If it is green, we need to change it's color to red, increment the
number of reds, and decrement the number of greens.