Get and check whether the neighbors are valid
Once we've toggled the current tile to red, we need to process the
current tile's neighbors. If we know what the current row and column
are, we can easily get the row and column of the neighbors by adding or
subtracting one to the appropriate variable.

Note that some of these cells will actually not be valid; for example,
if we are in the 0th row, trying to access the 0 - 1 = -1th row will be
invalid. We can check whether a variable is valid by checking the
appropriate condition... beware of off by one errors!

To make things easier, we're going to utilize the fact that JavaScript
variables are dynamically typed. Since we don't have to declare a
variable's type, we could assign either a jQuery DOM element or `null`
to a variable and let the JavaScript interpreter determine this at
runtime. Combining this with the JavaScript _ternary operator_,
`<condition> ? <value if true> : <value if false>`, we can very
conveniently check the edge conditions, get the neighbor if safe, and
evaluate to `null` if not.

Once we've done this, I'm going to stick all four variables into an
array with their corresponding coordinates so that we can conveniently
iterate over the array, and do the same processing for all four.