Use Array.forEach to iterate over each neighbor
This time, when iterating over each potential neighbor, we're going to
use a different style of JavaScript for loop called `Array.forEach`.
This is a function present on all JavaScript arrays that allows you to
loop over all the elements of a list without using indices.

The way this works is you specify a function to the `.forEach()` method.
This function takes as parameters the current element of the list, the
index of the current element, and the a reference to the array the
function was called on (in our example this isn't necessary because we
have access to the original array variable, but there are times when
this isn't the case). Since this function is a callback, it will be
evaluated once for each element of the array, and the appropriate values
will be substituted in.