Process each neighbor within Array.forEach
Within the forEach callback, we want to check to make sure that the
current neighbor we're processing exists, i.e. is not null. Since the
actual element we're iterating over is an object containing the
properties `tile`, `row`, and `col`, we need to check the
`$neighbor.tile` property to see if the element itself exists. Since
JavaScript values are all truthy, an object is only ever falsy if it's
null, we can simply use `if($neighbor.tile)` to check whether the
neighbor exists or not.

Now that we know whether the neighbor exists, we can do the appropriate
processing. Particularly, what we need to do depends on the color of the
current element. Let's get the color of the current neighbor using the
same method we did for the clicked tile (making care to draw the row and
column from the properties we set on the current `$neighbor` variable).

Now we can case on the color of the current element. If it's green,
we'll need to toggle it's color to blue and adjust the number of blues
and greens, and vice versa it the neighbor is blue.