Update scoreboard and check for winning conditions
The only thing left to do is update the scoreboard and check whether the
player has won the game.

If you recall the code to update the scoreboard from early, we will use
the same code now, but add an additional line to update reds (because
reds will have definitely changed this time).

If the number of reds is equal to the total number of tiles, that means
that every tile has been flipped, so the user has won. If the number of
greens is 0, then the user lost, because there aren't any tiles that can
be flipped. There's a JavaScript function `alert` that lets you display
a message in a dialog box, so we can use this to report a win or a loss.

It turns out that some boards are winnable and some boards aren't (if
you're interested you should try to prove it). Regardless of this, I
thought it'd be funny when I made this game and distributed it to my
fellow 251 classmates to tell them that a non-winnable board was in fact
winnable... it had some interesting consequences. Feel free to change
the winning and losing messages to something more "agreeable."

The final thing we'll take care of is reload the page for if the user
want to play again. We can do with with `location.reload()`, which will
take the current page's URL and reload that addresss.

That's it! You should be able to try everything out and have it work in
the browser. If you're still stuck, or you've got errors, flag down a
mentor or ask someone for help and we'd love to give you some hints to
fix everything up.

Thanks!